part 4: 4.21

+++ controllers/blogs.js
@@ -3,11 +3,14 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
+
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-    .populate('user', { username: 1, name: 1 })
-  response.json(blogs)
+    const blogs = await Blog.find({})
+        .populate('user', { username: 1, name: 1 })
+
+    response.json(blogs)
 })
+
 
 blogsRouter.post('/', async (request, response) => {
     const token = request.token
@@ -21,7 +24,10 @@ blogsRouter.post('/', async (request, response) => {
     let decodedToken
 
     try {
-        decodedToken = jwt.verify(token, process.env.SECRET)
+        decodedToken = jwt.verify(
+            token,
+            process.env.SECRET
+        )
     } catch (error) {
         return response.status(401).json({
             error: 'token invalid'
@@ -49,23 +55,69 @@ blogsRouter.post('/', async (request, response) => {
     response.status(201).json(result)
 })
 
+
 blogsRouter.delete('/:id', async (request, response) => {
+    if (!request.token) {
+        return response.status(401).json({
+            error: 'token missing'
+        })
+    }
 
-  const result = await Blog.findByIdAndDelete(request.params.id)
+    let decodedToken
 
-  response.status(204).end()
+    try {
+        decodedToken = jwt.verify(
+            request.token,
+            process.env.SECRET
+        )
+    } catch (error) {
+        return response.status(401).json({
+            error: 'token invalid'
+        })
+    }
 
+    if (!decodedToken.id) {
+        return response.status(401).json({
+            error: 'token invalid'
+        })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        return response.status(404).end()
+    }
+
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+        return response.status(401).json({
+            error: 'user not authorized to delete this blog'
+        })
+    }
+
+    await Blog.findByIdAndDelete(request.params.id)
+
+    response.status(204).end()
 })
 
 
 blogsRouter.put('/:id', async (request, response) => {
-  const blog = await Blog.findByIdAndUpdate(
-    request.params.id,
-    request.body,
-    { new: true }
-  )
+    const body = request.body
 
-  response.json(blog)
+    const updatedBlog = {
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes
+    }
+
+    const result = await Blog.findByIdAndUpdate(
+        request.params.id,
+        updatedBlog,
+        { new: true }
+    )
+
+    response.json(result)
 })
+
 
 module.exports = blogsRouter
